Add fakedata and split tests
