+++ test/_fakedata/README.md
@@ -1,0 +1,14 @@
+# Fake Data
+This directory is only used for testing purpose.
+
+## Data Structure
+### Files
+| Folder | File    | Data                                         |
+|--------|---------|----------------------------------------------|
+| bytes  | fn.bt   | b'fn\_0123456789abcdef'                      |
+| csv    | fn.csv  | key,item<br>0,fn\_0<br>1,fn\_1                     |
+| json   | fn.json | {'fn\_0': [{'fn\_01': 1}, {'fn\_02': 2}],'fn\_1': 1} |
+| txt    | fn.txt  | 'fn\_0123456789abcdef'                       |
+
+### Archives
+Each subfolder has corresponding archive files (uncompressed and compressed).
