+++ test/_fakedata/bytes/0.bt
@@ -1,0 +1,1 @@
+0_0123456789abcdef
