+++ test/_fakedata/bytes/1.bt
@@ -1,0 +1,1 @@
+1_0123456789abcdef
