+++ test/_fakedata/bytes/2.bt
@@ -1,0 +1,1 @@
+2_0123456789abcdef
