+++ test/_fakedata/csv/0.csv
@@ -1,0 +1,3 @@
+key,item
+0,0_0
+1,0_1
