+++ test/_fakedata/csv/1.csv
@@ -1,0 +1,3 @@
+key,item
+0,1_0
+1,1_1
