+++ test/_fakedata/csv/2.csv
@@ -1,0 +1,3 @@
+key,item
+0,2_0
+1,2_1
