+++ test/_fakedata/json/0.json
@@ -1,0 +1,1 @@
+{"0_0": [{"0_01": 1}, {"0_02": 2}], "0_1": 1}
