+++ test/_fakedata/json/1.json
@@ -1,0 +1,1 @@
+{"1_0": [{"1_01": 1}, {"1_02": 2}], "1_1": 1}
