+++ test/_fakedata/json/2.json
@@ -1,0 +1,1 @@
+{"2_0": [{"2_01": 1}, {"2_02": 2}], "2_1": 1}
