+++ test/_fakedata/txt/0.txt
@@ -1,0 +1,1 @@
+0_0123456789abcdef
